Moved calc_robustness to Experiment class
